# api_docs.yaml
openapi: 3.0.0
info:
    title: Micro Service API
    version: 1.0.0
servers:
    - url: data_fetcher:5000
    - url: data_manager:5001
    - url: job_runner:5002
paths:
...
    /api/internal/data_manager/add_tif:
        post:
            summary: Add files to sftp
            requestBody:
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                user_id:
                                    type: string
                                project_id:
                                    type: string
                                image_id:
                                    type: string
                                image:
                                    type: string
                                transform:
                                    type: object
                                    properties:
                                        x_coord:
                                            type: number
                                        y_coord:
                                            type: number
                                        x_meter:
                                            type: number
                                        y_meter:
                                            type: number
                                        x_rotation:
                                            type: number
                                        y_rotation:
                                            type: number

                                
            responses:
                200:
                    description: Successful Added
                400:
                    description: Bad request
    /api/internal/data_manager/user/add:
        post:
            summary: Add a new user
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                email:
                                    type: string
                                password:
                                    type: string
                                name:
                                    type: string
                                PLANET_API_KEY:
                                    type: string
            responses:
                '200':
                    description: User added successfully
                '400':
                    description: User was not added
    /api/internal/data_manager/user/verify:
        post:
            summary: Verify user credentials
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                email:
                                    type: string
                                password:
                                    type: string
            responses:
                '200':
                    description: User verified successfully
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    _id:
                                        type: string
                                    email:
                                        type: string
                                    name:
                                        type: string
                '400':
                    description: User verification failed
    /api/internal/data_manager/project/get:
        get:
            summary: Get user projects
            parameters:
                - in: header
                  name: user_id
                  required: true
                  schema:
                    type: string
            responses:
                '200':
                    description: Projects retrieved successfully
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    type: object
                '400':
                    description: Failed to retrieve projects
    /api/internal/data_manager/project/add:
        post:
            summary: Add a new project
            parameters:
                - in: header
                  name: user_id
                  required: true
                  schema:
                    type: string
            requestBody:
                required: true
                content:
                    application/json:
                        schema:
                            type: object
                            properties:
                                farm_name:
                                    type: string
                                aoi:
                                    type: array
                                    items:
                                        type: array
                                        items:
                                            type: number
                                crop:
                                    type: string
                                created_at:
                                    type: string
                                seeding_date:
                                    type: string
            responses:
                '200':
                    description: Project added successfully
                '400':
                    description: Failed to add project
...